build: enable -Werror=cast-qual + clean the 46 existing const-drop sites
Casts that drop `const` (or `volatile`) are the primitive that would
launder a sealed / .rodata table (module registry, dispatch vtables, the
manifest-derived allowlists) back into something writable. -Werror=cast-qual
makes any such cast in Hull's own source a hard compile error, so a future
change can't quietly reintroduce one. Vendored TUs keep their own relaxed
*_CFLAGS, so this is Hull-source-only.

The sweep found 41 sites in src/hull/** + include/hull/** and 5 more in
tests/. Notably NONE of them launder a sealed/rodata table — they are all
benign const-drops, which confirms the audit finding that every dispatch
table is already const-correct. The flag here is purely preventive.

The 46 sites, by fix:
- Owned-but-const-typed frees (the largest group: HlValue.value.s, cached
  SQL, app_dir, copied results). Two helpers added in hull/alloc.h:
  hl_free_const() / hl_alloc_free_const(), which launder through uintptr_t
  in one documented place. The const there is a typing artifact, not an
  immutability guarantee.
- In-place writes into a freshly-malloc'd buffer stored in a const field
  (worker deep-copy, table->kv). Rewritten to write through a `char *buf`
  local, then assign buf to the const field — no cast at all.
- qsort comparators: `(const char **)` -> `(const char *const *)` (the
  proper, fully-const cast — genuinely the correct fix).
- POSIX execvp (takes `char *const[]` but never modifies argv) and
  hl_cmd_modules argv: documented `(T)(uintptr_t)x`.
- Borrowed read-only data (image view, wasm buffer) and runtime image
  decode input: const-correct retype where clean, else documented launder.
- Test files: clean `(const void *)` for pointer comparisons; documented
  `(uintptr_t)` launder for the seal/tamper death tests that deliberately
  write to const memory.

Verified: clang `make` + `make test` 50/50 with -Werror=cast-qual (so the
tree is provably free of const-drops); gcc-14 -Werror=cast-qual clean on
the changed TUs (db/worker_db/migrate/tool) — the uintptr launder and
const-comparator patterns were pre-checked on gcc-14 since they must hold
on the Linux build too.

+++ src/hull/commands/check.c
@@ -14,6 +14,7 @@
 #include "hull/commands/verify.h"
 #include "hull/commands/modules.h"
 
+#include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -25,7 +26,7 @@ int hl_cmd_check(int argc, char **argv, const HlCommandEnv *env)
     fprintf(stderr, "[hull:check] validating manifest...\n");
     {
         const char *list_argv[3] = { "check", "list", env->app_dir };
-        int rc = hl_cmd_modules(3, (char **)list_argv, env);
+        int rc = hl_cmd_modules(3, (char **)(uintptr_t)list_argv, env);  /* hl_cmd_modules does not modify argv */
         if (rc != 0) {
             fprintf(stderr, "[hull:check] manifest validation failed\n");
             return rc;
@@ -40,7 +41,7 @@ int hl_cmd_check(int argc, char **argv, const HlCommandEnv *env)
     fprintf(stderr, "[hull:check] analyzing imports...\n");
     {
         const char *analyze_argv[3] = { "check", "analyze", env->app_dir };
-        int rc = hl_cmd_modules(3, (char **)analyze_argv, env);
+        int rc = hl_cmd_modules(3, (char **)(uintptr_t)analyze_argv, env);  /* hl_cmd_modules does not modify argv */
         if (rc != 0) {
             fprintf(stderr, "[hull:check] import analysis failed\n");
             return rc;
